+++ .github/workflows/run-notebooks.yml
@@ -5,7 +5,7 @@ jobs:
     runs-on: ubuntu-latest 
     strategy:
       matrix:
-        notebook: ['concepts/dataset/dataset-api.ipynb', 'concepts/model/model-api.ipynb', 'concepts/datastore/datastore-api.ipynb', 'concepts/compute/compute-instance-api.ipynb', 'concepts/compute/azureml-compute-api.ipynb', 'concepts/environment/environment-api.ipynb', 'concepts/workspace/workspace-api.ipynb', 'notebooks/fastai/train-pets-resnet34.ipynb', 'notebooks/fastai/train-mnist-mlproject.ipynb', 'notebooks/fastai/train-mnist-resnet18.ipynb', 'notebooks/triton/aks-gpu-deploy.ipynb', 'notebooks/tensorflow-v2/train-iris-nn.ipynb', 'notebooks/tensorflow-v2/train-mnist-nn.ipynb', 'notebooks/lightgbm/train-iris.ipynb', 'notebooks/pytorch/train-mnist-mlproject.ipynb', 'notebooks/pytorch/train-mnist-cnn.ipynb', 'notebooks/pytorch/deploy-mnist.ipynb', 'notebooks/xgboost/train-iris.ipynb', 'notebooks/sklearn/deploy-diabetes.ipynb', 'notebooks/sklearn/train-diabetes-mlproject.ipynb', 'notebooks/sklearn/train-diabetes-ridge.ipynb']
+        notebook: ['concepts/compute/azureml-compute-api.ipynb', 'concepts/compute/compute-instance-api.ipynb', 'concepts/dataset/dataset-api.ipynb', 'concepts/datastore/datastore-api.ipynb', 'concepts/environment/environment-api.ipynb', 'concepts/model/model-api.ipynb', 'concepts/workspace/workspace-api.ipynb', 'notebooks/fastai/train-mnist-mlproject.ipynb', 'notebooks/fastai/train-mnist-resnet18.ipynb', 'notebooks/fastai/train-pets-resnet34.ipynb', 'notebooks/lightgbm/train-iris.ipynb', 'notebooks/pytorch/deploy-mnist.ipynb', 'notebooks/pytorch/train-mnist-cnn.ipynb', 'notebooks/pytorch/train-mnist-mlproject.ipynb', 'notebooks/sklearn/deploy-diabetes.ipynb', 'notebooks/sklearn/train-diabetes-mlproject.ipynb', 'notebooks/sklearn/train-diabetes-ridge.ipynb', 'notebooks/tensorflow-v2/train-iris-nn.ipynb', 'notebooks/tensorflow-v2/train-mnist-nn.ipynb', 'notebooks/triton/deploy-densenet-local.ipynb', 'notebooks/xgboost/train-iris.ipynb']
     steps:
     - name: check out repo
       uses: actions/checkout@v2
