--- notebooks/triton/deploymentconfig.json
@@ -1,8 +1,0 @@
-{
-    "computeType": "aks",
-    "containerResourceRequirements":{
-        "cpu": 1,
-        "memoryInGB": 4
-    },
-    "gpuCores": 1
-}
