--- notebooks/triton/inference-config.json
@@ -1,4 +1,0 @@
-{
-    "entryScript": "score_densenet.py",
-    "sourceDirectory": "../scripts"
-}
